fix bugs time message

+++ parse.php
@@ -4,25 +4,13 @@ include("simple_html_dom.php");
 //Remove Warnings
 error_reporting(E_ERROR | E_PARSE);
 date_default_timezone_set('Europe/Paris');
-echo date("d m y G i s");
+echo date("d/m/y G:i:s");
 
 // Main vars
 $urls = [];
 
-/*
-$contentHTML = file_get_html('http://iutsa.unice.fr/gpushow2/?dept=RT&interactive&date=plus&filiere=RT1,RT1M,RT2,LPRT-RSFS');
-
-foreach (parseHTML($contentHTML) as $class) {
-    echo 'Un ' . $class['typeClass'] . ' est prévu à ' . $class['time'] . ' en ' . $class['group'] . ' à la salle "' . $class['room'] . '". Le professeur sera ' . $class['teacher'] . ' et le module enseigné est ' . $class['module'] . '.<br>';
-}
-*/
-
-$logFile = fopen(dirname(__FILE__) . "/logParse.log", "a") or die("Unable to open file!");
-        fwrite($logFile, 'done=true');
-        fclose($logFile);
-        
-$beginTime = mktime(19, 0, 0, date('m'), date("d") , date("Y"));
-$endTime = mktime(20, 0, 0, date('m'), date("d") , date("Y"));
+$beginTime = mktime(20, 0, 0, date('m'), date("d") , date("Y"));
+$endTime = mktime(21, 0, 0, date('m'), date("d") , date("Y"));
 
 if(time() > $beginTime && time() < $endTime) {
     if(!file_exists( dirname(__FILE__) . '/alreadyDone.file')) {
@@ -38,10 +26,6 @@ if(time() > $beginTime && time() < $endTime) {
         
         echo $message;
         
-        $logFile = fopen(dirname(__FILE__) . "/logParse.log", "a") or die("Unable to open file!");
-        fwrite($logFile, $message . "\n");
-        fclose($logFile);
-        
         sendMessages($message);
         
         $myfile = fopen(dirname(__FILE__) . "/alreadyDone.file", "w") or die("Unable to open file!");
@@ -50,30 +34,30 @@ if(time() > $beginTime && time() < $endTime) {
     }
     else {
         echo 'alreadyDone';
-        
-        unlink(dirname(__FILE__) . '/alreadyDone.file');
     }
 }
 else {
+    if(file_exists( dirname(__FILE__) . '/alreadyDone.file')) {
+	unlink(dirname(__FILE__) . '/alreadyDone.file');
+    }
+
     $contentHTML = getTimeTable(0);
     $classes = parseHTML($contentHTML);
     
     foreach ($classes as $class) {
         $message =  'Un ' . str_replace("\r\n", "", $class['typeClass']) . ' est prévu à ' . $class['time'] . ' en ' . $class['group'] . ' à la salle "' . $class['room'] . '". Le professeur sera ' . $class['teacher'] . ' et le module enseigné est ' . str_replace(' ', '', $class['module']) . '.';
         
-        echo $class['remainingTime'] . ' - ' . $message . '<br>';
+        echo ' - ' . $class['remainingTime'] . ' - ' . $message . '<br>';
         
-        if($class['remainingTime'] < 10 && $class['remainingTime'] != 'DONE' && $class['group'] != 'other group') {
+        if($class['remainingTime'] < 7 && $class['remainingTime'] > 2 && $class['remainingTime'] != 'DONE' && $class['group'] != 'other group') {
             sendMessages($message);
         }
-        
-        $logFile = fopen(dirname(__FILE__) . "/logParse.log", "a") or die("Unable to open file!");
-        fwrite($logFile, $message . "\n");
-        fclose($logFile);
     }
 }
 
 function parseHTML($html) {
+    $tabTimes[] = '16:00';
+    
     foreach(array_slice($html->find('div.edt3'), 0, 22) as $div) {
         $tabTimes[] = (sizeof(explode(':', $div->plaintext)) > 1) ? $div->plaintext : ($div->plaintext . ':00');
     }
@@ -95,6 +79,14 @@ function parseHTML($html) {
         $explodedTimeCours = explode(':', $time);
         $timeClass = mktime($explodedTimeCours[0], $explodedTimeCours[1],0, date('m'), date("d") , date("Y"));
         $remainingTime = getdate($timeClass - time())['minutes'];
+        
+        if (($timeClass - time()) > 0)
+        {
+            $remainingTimeText = $remainingTime;
+        }
+        else {
+            $remainingTimeText = 'DONE';
+        }
         
         //INFORMATIONS
         $child = $div->first_child();
@@ -122,23 +114,6 @@ function parseHTML($html) {
                 else {
                     $group = 'other group';
                 }
-            }
-        
-            if($group != 'other group') {
-                //$smsText = 'Un ' . $typeClass . ' est prévu à ' . $time . ' en ' . $group . ' à la salle "' . $room . '". Le professeur sera ' . $teacher . ' et le module enseigné ' . $module . '.';
-                //echo $smsText;
-                
-                if (($timeClass - time()) > 0)
-                {
-                    if (($timeClass - time()) <= 600)
-                    {
-                        $remainingTimeText = $remainingTime['minutes'];
-                    }
-                }
-                else {
-                    $remainingTimeText = 'DONE';
-                }
-    
             }
             
             $tabClasses[] = ['typeClass' => $typeClass, 'time' => $time, 'remainingTime' => $remainingTimeText, 'group' => $group, 'room' => $room, 'teacher' => $teacher, 'module' => $module];
@@ -199,3 +174,5 @@ function getTimeTable($daysToLoad = 0)
     return str_get_html($html);
     }
 }
+
+print_r($tabTimes);
